requirements.txt
pyautogui
mss
rich
numpy
keyboard
scipy
Pillow
pytesseract
